requirements.txt
selenium==4.10.0
webdriver-manager==3.8.6
httpx==0.24.1
beautifulsoup4==4.12.2
openai==1.6.1
python-dotenv==1.0.0 